+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Add `AbstractServerTest.addPostTestSqlAction(...)` method to help reversing destructive actions after a test has completed.
+
 ### [v0.100](https://github.com/realityforge/guiceyloops/tree/v0.100) (2019-04-26)
 [Full Changelog](https://github.com/realityforge/guiceyloops/compare/v0.99...v0.100)